requirements.txt
jpegio
pillow
pandas
numpy
streamlit
